# src/main/kotlin/com/wutsi/platform/account/endpoint/UpdateAccountAttributeController.kt
package com.wutsi.platform.account.endpoint

import com.wutsi.platform.account.`delegate`.UpdateAccountAttributeDelegate
import com.wutsi.platform.account.dto.UpdateAccountAttributeRequest
import org.springframework.security.access.prepost.PreAuthorize
import org.springframework.web.bind.`annotation`.PathVariable
import org.springframework.web.bind.`annotation`.PostMapping
import org.springframework.web.bind.`annotation`.RequestBody
import org.springframework.web.bind.`annotation`.RestController
import javax.validation.Valid
import javax.validation.constraints.NotBlank
import javax.validation.constraints.NotNull
import kotlin.Long
import kotlin.String

@RestController
public class UpdateAccountAttributeController(
    private val `delegate`: UpdateAccountAttributeDelegate
) {
    @PostMapping("/v1/accounts/{id}/attributes/{name}")
    @PreAuthorize(value = "hasAuthority('user-manage')")
    public fun invoke(
        @PathVariable(name = "id") @NotNull id: Long,
        @PathVariable(name = "name") @NotBlank name: String,
        @Valid @RequestBody request: UpdateAccountAttributeRequest
    ) {
        delegate.invoke(id, name, request)
    }
}